clean and normalize tests
