+++ tests/MaintenanceTests/Provider/ConfigProviderTest.php
@@ -24,164 +24,15 @@ class ConfigProviderTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::attach
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::detach
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::setMessage
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::getMessage
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::setBlock
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::isBlocked
-     */
-    public function testAttachDetach()
-    {
-        $eventManager = $this->getMock('Zend\\EventManager\\EventManagerInterface');
-        $callbackMock = $this->getMock('Zend\\Stdlib\\CallbackHandler', array(), array(), '', false);
-
-        $this->provider->setMessage('custom message');
-        $this->assertEquals('custom message', $this->provider->getMessage());
-
-        $this->provider->setBlock(false);
-        $this->assertFalse($this->provider->isBlocked());
-
-        $eventManager
-            ->expects($this->once())
-            ->method('attach')
-            ->with()
-            ->will($this->returnValue($callbackMock));
-        $this->provider->attach($eventManager);
-
-        $eventManager
-            ->expects($this->once())
-            ->method('detach')
-            ->with($callbackMock)
-            ->will($this->returnValue(true));
-        $this->provider->detach($eventManager);
-    }
-
-    /**
      * @covers Jgut\Zf\Maintenance\Provider\ConfigProvider::setActive
      * @covers Jgut\Zf\Maintenance\Provider\ConfigProvider::isActive
      */
-    public function testIsActive()
+    public function testMutatorsAccessors()
     {
+        $this->assertFalse($this->provider->isActive());
+
         $this->provider->setActive(true);
 
         $this->assertTrue($this->provider->isActive());
-    }
-
-    /**
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::onRoute
-     */
-    public function testNotIsActive()
-    {
-        $this->provider->setActive(true);
-        $this->provider->setBlock(false);
-
-        $event = $this->getMock('Zend\\Mvc\\MvcEvent', array(), array(), '', false);
-
-        $this->assertNull($this->provider->onRoute($event));
-    }
-
-    /**
-     * @covers Jgut\Zf\Maintenance\Provider\ConfigProvider::setActive
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::onRoute
-     */
-    public function testNoRouteMatch()
-    {
-        $this->provider->setActive(true);
-
-        $event = $this->getMock('Zend\\Mvc\\MvcEvent', array(), array(), '', false);
-        $event->expects($this->any())->method('getRouteMatch')->will($this->returnValue(false));
-
-        $event->expects($this->never())->method('setError');
-        $this->provider->onRoute($event);
-    }
-
-    /**
-     * @covers Jgut\Zf\Maintenance\Provider\ConfigProvider::setActive
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::onRoute
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::isExcluded
-     */
-    public function testIsExcluded()
-    {
-        $exclusion = $this->getMock('Jgut\\Zf\\Maintenance\\Exclusion\\IpExclusion', array(), array(), '', false);
-        $exclusion->expects($this->once())->method('isExcluded')->will($this->returnValue(true));
-
-        $options = $this->getMock('Jgut\\Zf\\Maintenance\\Options\\ModuleOptions', array(), array(), '', false);
-        $options->expects($this->once())->method('getExclusions')->will(
-            $this->returnValue(array(
-                'ZfMaintenanceRouteExclusion' => '',
-                'ZfMaintenanceIpExclusion'    => '',
-            ))
-        );
-
-        $serviceManager = $this->getMock('Zend\\ServiceManager\\ServiceManager', array(), array(), '', false);
-        $serviceManager->expects($this->any())->method('has')->will(
-            $this->returnCallback(
-                function () {
-                    $args = array(
-                        'ZfMaintenanceOptions',
-                        'ZfMaintenanceIpExclusion',
-                    );
-                    return in_array(func_get_arg(0), $args);
-                }
-            )
-        );
-        $serviceManager->expects($this->any())->method('get')->will(
-            $this->returnCallback(
-                function () use ($options, $exclusion) {
-                    $args = array(
-                        'ZfMaintenanceOptions'     => $options,
-                        'ZfMaintenanceIpExclusion' => $exclusion
-                    );
-                    return $args[func_get_arg(0)];
-                }
-            )
-        );
-
-        $application = $this->getMock('Zend\\Mvc\\Application', array(), array(), '', false);
-        $application->expects($this->once())->method('getServiceManager')->will($this->returnValue($serviceManager));
-
-        $routeMatch = $this->getMock('Zend\\Mvc\\Router\\RouteMatch', array(), array(), '', false);
-
-        $event = $this->getMock('Zend\\Mvc\\MvcEvent', array(), array(), '', false);
-        $event->expects($this->once())->method('getRouteMatch')->will($this->returnValue($routeMatch));
-        $event->expects($this->any())->method('getApplication')->will($this->returnValue($application));
-
-        $event->expects($this->never())->method('setError');
-
-        $this->provider->setActive(true);
-
-        $this->provider->onRoute($event);
-    }
-
-    /**
-     * @covers Jgut\Zf\Maintenance\Provider\ConfigProvider::setActive
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::onRoute
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::isExcluded
-     */
-    public function testIsNotExcluded()
-    {
-        $eventManager = $this->getMock('Zend\\EventManager\\EventManagerInterface');
-        $eventManager->expects($this->once())->method('trigger');
-
-        $options = $this->getMock('Jgut\\Zf\\Maintenance\\Options\\ModuleOptions', array(), array(), '', false);
-        $options->expects($this->once())->method('getExclusions')->will($this->returnValue(array()));
-
-        $serviceManager = $this->getMock('Zend\\ServiceManager\\ServiceManager', array(), array(), '', false);
-        $serviceManager->expects($this->once())->method('get')->will($this->returnValue($options));
-
-        $application = $this->getMock('Zend\\Mvc\\Application', array(), array(), '', false);
-        $application->expects($this->once())->method('getServiceManager')->will($this->returnValue($serviceManager));
-        $application->expects($this->once())->method('getEventManager')->will($this->returnValue($eventManager));
-
-        $routeMatch = $this->getMock('Zend\\Mvc\\Router\\RouteMatch', array(), array(), '', false);
-
-        $event = $this->getMock('Zend\\Mvc\\MvcEvent', array(), array(), '', false);
-        $event->expects($this->once())->method('getRouteMatch')->will($this->returnValue($routeMatch));
-        $event->expects($this->any())->method('getApplication')->will($this->returnValue($application));
-
-        $this->provider->setActive(true);
-
-        $this->provider->onRoute($event);
     }
 }
