+++ tests/MaintenanceTests/Provider/ConfigScheduledProviderTest.php
@@ -50,7 +50,7 @@ class ConfigScheduledProviderTest extends PHPUnit_Framework_TestCase
      */
     public function testTimes()
     {
-        $start = new \DateTime('now');
+        $start = new \DateTime();
         $start->add(new \DateInterval('P1D'));
 
         $this->provider->setStart($start);
@@ -61,7 +61,7 @@ class ConfigScheduledProviderTest extends PHPUnit_Framework_TestCase
 
         $this->provider = new ConfigScheduledProvider();
 
-        $end = new \DateTime('now');
+        $end = new \DateTime();
         $end->add(new \DateInterval('P2D'));
 
         $this->provider->setEnd($end);
@@ -77,7 +77,7 @@ class ConfigScheduledProviderTest extends PHPUnit_Framework_TestCase
 
         $this->provider = new ConfigScheduledProvider();
 
-        $end = new \DateTime('now');
+        $end = new \DateTime();
         $end->sub(new \DateInterval('P2D'));
 
         $this->provider->setEnd($end);
@@ -92,10 +92,10 @@ class ConfigScheduledProviderTest extends PHPUnit_Framework_TestCase
      */
     public function testInvalidEndDate()
     {
-        $start = new \DateTime('now');
+        $start = new \DateTime();
         $this->provider->setStart($start);
 
-        $end = new \DateTime('now');
+        $end = new \DateTime();
         $end->sub(new \DateInterval('P2D'));
         $this->provider->setEnd($end);
     }
@@ -105,10 +105,10 @@ class ConfigScheduledProviderTest extends PHPUnit_Framework_TestCase
      */
     public function testInvalidStartDate()
     {
-        $end = new \DateTime('now');
+        $end = new \DateTime();
         $this->provider->setEnd($end);
 
-        $start = new \DateTime('now');
+        $start = new \DateTime();
         $start->add(new \DateInterval('P2D'));
         $this->provider->setStart($start);
     }
@@ -116,153 +116,40 @@ class ConfigScheduledProviderTest extends PHPUnit_Framework_TestCase
     /**
      * @covers Jgut\Zf\Maintenance\Provider\ConfigScheduledProvider::setStart
      * @covers Jgut\Zf\Maintenance\Provider\ConfigScheduledProvider::setEnd
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::onRoute
+     * @covers Jgut\Zf\Maintenance\Provider\ConfigScheduledProvider::isActive
      */
     public function testBeforeTime()
     {
-        $start = new \DateTime('now');
+        $start = new \DateTime();
         $start->add(new \DateInterval('P1D'));
         $this->provider->setStart($start);
 
-        $event = $this->getMock('Zend\\Mvc\\MvcEvent', array(), array(), '', false);
+        $this->assertFalse($this->provider->isActive());
 
-        $this->assertNull($this->provider->onRoute($event));
-
-        $end = new \DateTime('now');
+        $end = new \DateTime();
         $end->add(new \DateInterval('P2D'));
         $this->provider->setStart($end);
 
-        $this->assertNull($this->provider->onRoute($event));
+        $this->assertFalse($this->provider->isActive());
     }
 
     /**
      * @covers Jgut\Zf\Maintenance\Provider\ConfigScheduledProvider::setStart
      * @covers Jgut\Zf\Maintenance\Provider\ConfigScheduledProvider::setEnd
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::onRoute
+     * @covers Jgut\Zf\Maintenance\Provider\ConfigScheduledProvider::isActive
      */
     public function testAfterTime()
     {
-        $end = new \DateTime('now');
+        $end = new \DateTime();
         $end->sub(new \DateInterval('P1D'));
         $this->provider->setEnd($end);
 
-        $event = $this->getMock('Zend\\Mvc\\MvcEvent', array(), array(), '', false);
+        $this->assertFalse($this->provider->isActive());
 
-        $this->assertNull($this->provider->onRoute($event));
-
-        $start = new \DateTime('now');
+        $start = new \DateTime();
         $start->sub(new \DateInterval('P2D'));
         $this->provider->setStart($start);
 
-        $this->assertNull($this->provider->onRoute($event));
-    }
-
-    /**
-     * @covers Jgut\Zf\Maintenance\Provider\ConfigScheduledProvider::setStart
-     * @covers Jgut\Zf\Maintenance\Provider\ConfigScheduledProvider::setEnd
-     * @covers Jgut\Zf\Maintenance\Provider\ConfigScheduledProvider::onRoute
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::onRoute
-     */
-    public function testNoRouteMatch()
-    {
-        $start = new \DateTime('now');
-        $start->sub(new \DateInterval('P1D'));
-        $this->provider->setStart($start);
-
-        $end = new \DateTime('now');
-        $end->add(new \DateInterval('P1D'));
-        $this->provider->setEnd($end);
-
-        $event = $this->getMock('Zend\\Mvc\\MvcEvent', array(), array(), '', false);
-        $event->expects($this->any())->method('getRouteMatch')->will($this->returnValue(false));
-
-        $event->expects($this->never())->method('setError');
-        $this->provider->onRoute($event);
-    }
-
-    /**
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::onRoute
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::isExcluded
-     */
-    public function testIsExcluded()
-    {
-        $exclusion = $this->getMock('Jgut\\Zf\\Maintenance\\Exclusion\\IpExclusion', array(), array(), '', false);
-        $exclusion->expects($this->once())->method('isExcluded')->will($this->returnValue(true));
-
-        $options = $this->getMock('Jgut\\Zf\\Maintenance\\Options\\ModuleOptions', array(), array(), '', false);
-        $options->expects($this->once())->method('getExclusions')->will(
-            $this->returnValue(array('ZfMaintenanceIpExclusion' => ''))
-        );
-
-        $serviceManager = $this->getMock('Zend\\ServiceManager\\ServiceManager', array(), array(), '', false);
-        $serviceManager->expects($this->any())->method('has')->will($this->returnValue(true));
-        $serviceManager->expects($this->any())->method('get')->will(
-            $this->returnCallback(
-                function () use ($options, $exclusion) {
-                    $args = array(
-                        'ZfMaintenanceOptions'     => $options,
-                        'ZfMaintenanceIpExclusion' => $exclusion
-                    );
-                    return $args[func_get_arg(0)];
-                }
-            )
-        );
-
-        $application = $this->getMock('Zend\\Mvc\\Application', array(), array(), '', false);
-        $application->expects($this->once())->method('getServiceManager')->will($this->returnValue($serviceManager));
-
-        $routeMatch = $this->getMock('Zend\\Mvc\\Router\\RouteMatch', array(), array(), '', false);
-
-        $event = $this->getMock('Zend\\Mvc\\MvcEvent', array(), array(), '', false);
-        $event->expects($this->once())->method('getRouteMatch')->will($this->returnValue($routeMatch));
-        $event->expects($this->any())->method('getApplication')->will($this->returnValue($application));
-
-        $event->expects($this->never())->method('setError');
-
-        $start = new \DateTime('now');
-        $start->sub(new \DateInterval('P1D'));
-        $this->provider->setStart($start);
-
-        $end = new \DateTime('now');
-        $end->add(new \DateInterval('P1D'));
-        $this->provider->setEnd($end);
-
-        $this->provider->onRoute($event);
-    }
-
-    /**
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::onRoute
-     * @covers Jgut\Zf\Maintenance\Provider\AbstractProvider::isExcluded
-     */
-    public function testIsNotExcluded()
-    {
-        $eventManager = $this->getMock('Zend\\EventManager\\EventManagerInterface');
-        $eventManager->expects($this->once())->method('trigger');
-
-        $options = $this->getMock('Jgut\\Zf\\Maintenance\\Options\\ModuleOptions', array(), array(), '', false);
-        $options->expects($this->once())->method('getExclusions')->will($this->returnValue(array()));
-
-        $serviceManager = $this->getMock('Zend\\ServiceManager\\ServiceManager', array(), array(), '', false);
-        $serviceManager->expects($this->once())->method('get')->will($this->returnValue($options));
-
-        $application = $this->getMock('Zend\\Mvc\\Application', array(), array(), '', false);
-        $application->expects($this->once())->method('getServiceManager')->will($this->returnValue($serviceManager));
-        $application->expects($this->once())->method('getEventManager')->will($this->returnValue($eventManager));
-
-        $routeMatch = $this->getMock('Zend\\Mvc\\Router\\RouteMatch', array(), array(), '', false);
-
-        $event = $this->getMock('Zend\\Mvc\\MvcEvent', array(), array(), '', false);
-        $event->expects($this->once())->method('getRouteMatch')->will($this->returnValue($routeMatch));
-        $event->expects($this->any())->method('getApplication')->will($this->returnValue($application));
-
-        $start = new \DateTime('now');
-        $start->sub(new \DateInterval('P1D'));
-        $this->provider->setStart($start);
-
-        $end = new \DateTime('now');
-        $end->add(new \DateInterval('P1D'));
-        $this->provider->setEnd($end);
-
-        $this->provider->onRoute($event);
+        $this->assertFalse($this->provider->isActive());
     }
 }
