+++ tests/MaintenanceTests/Provider/CrontabProviderTest.php
@@ -64,46 +64,54 @@ class CrontabProviderTest extends PHPUnit_Framework_TestCase
     /**
      * @covers Jgut\Zf\Maintenance\Provider\CrontabProvider::isActive
      */
-    public function testActive()
+    public function testNotIsActive()
+    {
+        $time   = new \DateTime();
+        $minute = $time->format('i');
+        $hour   = $time->format('H');
+        $day    = $time->format('d');
+        $month  = $time->format('m');
+
+        $this->provider->setInterval('PT10S');
+
+        $this->provider->setExpression(sprintf('%s * * * *', $minute - 10 > -1 ? $minute - 10 : 50));
+        $this->assertFalse($this->provider->isActive());
+
+        $this->provider->setExpression(sprintf('* %s * * *', $hour - 1 > -1 ? $hour - 1 : 23));
+        $this->assertFalse($this->provider->isActive());
+
+        $this->provider->setExpression(sprintf('* * %s * *', $day - 1 > -1 ? $day - 1 : 28));
+        $this->assertFalse($this->provider->isActive());
+
+        $this->provider->setExpression(sprintf('* * * %s *', $month - 1 > -1 ? $month - 1 : 12));
+        $this->assertFalse($this->provider->isActive());
+    }
+
+    /**
+     * @covers Jgut\Zf\Maintenance\Provider\CrontabProvider::isActive
+     */
+    public function testIsActive()
     {
         $time = new \DateTime();
+        $minute = $time->format('i');
+        $hour   = $time->format('H');
+        $day    = $time->format('d');
+        $month  = $time->format('m');
 
         $this->provider->setInterval('PT1M');
-        $this->provider->setExpression($time->format('i') . ' * * * *');
+        $this->provider->setExpression(sprintf('%s * * * *', $minute));
         $this->assertTrue($this->provider->isActive());
 
         $this->provider->setInterval('PT1H');
-        $this->provider->setExpression('* ' . $time->format('H') . ' * * *');
+        $this->provider->setExpression(sprintf('* %s * * *', $hour));
         $this->assertTrue($this->provider->isActive());
 
         $this->provider->setInterval('P1D');
-        $this->provider->setExpression('* * ' . $time->format('d') . ' * *');
+        $this->provider->setExpression(sprintf('* * %s * *', $day));
         $this->assertTrue($this->provider->isActive());
 
         $this->provider->setInterval('P1M');
-        $this->provider->setExpression('* * * ' . $time->format('m') . ' *');
-        $this->assertTrue($this->provider->isActive());
-
-        $this->provider->setInterval('PT1M');
-        $this->provider->setExpression($time->format('i') . ' ' . $time->format('H') . ' * * *');
-        $this->assertTrue($this->provider->isActive());
-
-        $this->provider->setExpression(
-            $time->format('i') . ' ' . $time->format('H') . ' ' . $time->format('d') . ' * *'
-        );
-        $this->assertTrue($this->provider->isActive());
-
-        $this->provider->setExpression(
-            $time->format('i') . ' ' . $time->format('H') . ' ' . $time->format('d') . ' ' . $time->format('m') . ' *'
-        );
-        $this->assertTrue($this->provider->isActive());
-
-        $this->provider->setExpression(
-            '* ' . $time->format('H') . ' ' . $time->format('d') . ' ' . $time->format('m') . ' *'
-        );
-        $this->assertTrue($this->provider->isActive());
-
-        $this->provider->setExpression('* * ' . $time->format('d') . ' ' . $time->format('m') . ' *');
+        $this->provider->setExpression(sprintf('* * * %s *', $month));
         $this->assertTrue($this->provider->isActive());
     }
